requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
numpy<2
httpx
transformers==4.39.0
torch
python-dotenv
python-dotenv
pandas
dill
xxhash
fsspec[http]<=2025.9.0,>=2023.1.0

# 데이터베이스 연동
SQLAlchemy
psycopg2-binary
redis

# LLM
openai
google-generativeai

# RAG
chromadb
sentence-transformers==2.7.0

